Add files via upload
CSIRO Adjusted Sea Level:
Commonwealth Scientific and Industrial Research Organization Adjusted Sea Level is a dataset that provides information on changes in sea level due to factors such as climate change and natural variations.
Lower Error Bound:
The lower limit of uncertainty or margin of error associated with the sea level measurements. Represents the range within which the true sea level measurement is likely to fall.
Upper Error Bound:
The upper limit of uncertainty or margin of error associated with the sea level measurements. Similar to the Lower Error Bound, the Upper Error Bound represents the range within which the true sea level measurement is likely to fall.